requirements.txt
altair==5.5.0
attrs==24.2.0
blinker==1.9.0
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
filelock==3.16.1
folium==0.18.0
fontawesome==5.10.1.post1
fsspec==2024.10.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
googlemaps==4.10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.15.1
networkx==3.4.2
numpy==2.1.3
osmnx==2.0.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
polyline==2.0.2
protobuf==5.29.0
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
streamlit==1.40.2
streamlit_folium==0.23.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xyzservices==2024.9.0
